Changed m_seqRefreshRate and associated usage from an int to a float. This handles when a sequence file has an uncommon Step Time. This causes the result of "1000 / Step Time" to not be an integer. Discovered with a Step Time of 17ms which caused the audio and lights to not stay in sync. At the end of the audio, the lights continued for several seconds before completing.

+++ src/channeloutput/channeloutputthread.c
@@ -44,7 +44,7 @@
 #include "settings.h"
 
 /* used by external sync code */
-int   RefreshRate = 20;
+float   RefreshRate = 20;
 int   DefaultLightDelay = 0;
 int   LightDelay = 0;
 volatile int FrameSkip = 0;
@@ -292,12 +292,12 @@ void *RunChannelOutputThread(void *data)
 /*
  * Set the step time
  */
-void SetChannelOutputRefreshRate(int rate)
+void SetChannelOutputRefreshRate(float rate)
 {
 	RefreshRate = rate;
 	DefaultLightDelay = 1000000 / RefreshRate;
 }
-int GetChannelOutputRefreshRate() {
+float GetChannelOutputRefreshRate() {
     return RefreshRate;
 }
